move new schema migration code to a different mod
